+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Magento 2 Goomento Page Builder
 
+## 0.3.9: Fixed mirror bugs
+- Fixed mirror bugs
+- Added widget preview
+- Updates revisions control
+
 ## 0.3.8: Optimized speed loading
 - Optimized speed loading
 - Added Builder Assistance to specify page